+++ crates/core/machine/src/runtime/syscall.rs
@@ -150,6 +150,9 @@ pub enum SyscallCode {
 
     /// Executes the `BN254_FP2_MUL` precompile.
     BN254_FP2_MUL = 0x00_01_01_2B,
+
+    /// Executes the `POSEIDON2_PERMUTE` precompile.
+    POSEIDON2_PERMUTE = 0x00_01_01_2F,
 }
 
 impl SyscallCode {
@@ -194,6 +197,7 @@ impl SyscallCode {
             0x00_01_01_2A => SyscallCode::BN254_FP2_SUB,
             0x00_01_01_2B => SyscallCode::BN254_FP2_MUL,
             0x00_00_01_1C => SyscallCode::BLS12381_DECOMPRESS,
+            0x00_01_01_2F => SyscallCode::POSEIDON2_PERMUTE,
             _ => panic!("invalid syscall number: {}", value),
         }
     }
