+++ crates/primitives/src/lib.rs
@@ -1,1167 +1,9 @@
-//! sp1-primitives contains types and functions that are used in both sp1-core and sp1-zkvm.
-//! Because it is imported in the zkvm entrypoint, it should be kept minimal.
-
-use lazy_static::lazy_static;
-use p3_baby_bear::{BabyBear, DiffusionMatrixBabyBear};
-use p3_field::AbstractField;
-use p3_poseidon2::{Poseidon2, Poseidon2ExternalMatrixGeneral};
-
 pub mod consts;
 pub mod io;
+pub mod poseidon2;
 pub mod types;
 
-lazy_static! {
-    // These constants are created by a RNG.
-
-    // This will be compatible with a poseidon2 permutation config with
-    // a state width of 16 and total rounds (both full and partial) of 30.
-    pub static ref RC_16_30: [[BabyBear; 16]; 30] = [
-        [
-            BabyBear::from_wrapped_u32(2110014213),
-            BabyBear::from_wrapped_u32(3964964605),
-            BabyBear::from_wrapped_u32(2190662774),
-            BabyBear::from_wrapped_u32(2732996483),
-            BabyBear::from_wrapped_u32(640767983),
-            BabyBear::from_wrapped_u32(3403899136),
-            BabyBear::from_wrapped_u32(1716033721),
-            BabyBear::from_wrapped_u32(1606702601),
-            BabyBear::from_wrapped_u32(3759873288),
-            BabyBear::from_wrapped_u32(1466015491),
-            BabyBear::from_wrapped_u32(1498308946),
-            BabyBear::from_wrapped_u32(2844375094),
-            BabyBear::from_wrapped_u32(3042463841),
-            BabyBear::from_wrapped_u32(1969905919),
-            BabyBear::from_wrapped_u32(4109944726),
-            BabyBear::from_wrapped_u32(3925048366),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3706859504),
-            BabyBear::from_wrapped_u32(759122502),
-            BabyBear::from_wrapped_u32(3167665446),
-            BabyBear::from_wrapped_u32(1131812921),
-            BabyBear::from_wrapped_u32(1080754908),
-            BabyBear::from_wrapped_u32(4080114493),
-            BabyBear::from_wrapped_u32(893583089),
-            BabyBear::from_wrapped_u32(2019677373),
-            BabyBear::from_wrapped_u32(3128604556),
-            BabyBear::from_wrapped_u32(580640471),
-            BabyBear::from_wrapped_u32(3277620260),
-            BabyBear::from_wrapped_u32(842931656),
-            BabyBear::from_wrapped_u32(548879852),
-            BabyBear::from_wrapped_u32(3608554714),
-            BabyBear::from_wrapped_u32(3575647916),
-            BabyBear::from_wrapped_u32(81826002),
-        ],
-        [
-            BabyBear::from_wrapped_u32(4289086263),
-            BabyBear::from_wrapped_u32(1563933798),
-            BabyBear::from_wrapped_u32(1440025885),
-            BabyBear::from_wrapped_u32(184445025),
-            BabyBear::from_wrapped_u32(2598651360),
-            BabyBear::from_wrapped_u32(1396647410),
-            BabyBear::from_wrapped_u32(1575877922),
-            BabyBear::from_wrapped_u32(3303853401),
-            BabyBear::from_wrapped_u32(137125468),
-            BabyBear::from_wrapped_u32(765010148),
-            BabyBear::from_wrapped_u32(633675867),
-            BabyBear::from_wrapped_u32(2037803363),
-            BabyBear::from_wrapped_u32(2573389828),
-            BabyBear::from_wrapped_u32(1895729703),
-            BabyBear::from_wrapped_u32(541515871),
-            BabyBear::from_wrapped_u32(1783382863),
-        ],
-        [
-            BabyBear::from_wrapped_u32(2641856484),
-            BabyBear::from_wrapped_u32(3035743342),
-            BabyBear::from_wrapped_u32(3672796326),
-            BabyBear::from_wrapped_u32(245668751),
-            BabyBear::from_wrapped_u32(2025460432),
-            BabyBear::from_wrapped_u32(201609705),
-            BabyBear::from_wrapped_u32(286217151),
-            BabyBear::from_wrapped_u32(4093475563),
-            BabyBear::from_wrapped_u32(2519572182),
-            BabyBear::from_wrapped_u32(3080699870),
-            BabyBear::from_wrapped_u32(2762001832),
-            BabyBear::from_wrapped_u32(1244250808),
-            BabyBear::from_wrapped_u32(606038199),
-            BabyBear::from_wrapped_u32(3182740831),
-            BabyBear::from_wrapped_u32(73007766),
-            BabyBear::from_wrapped_u32(2572204153),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1196780786),
-            BabyBear::from_wrapped_u32(3447394443),
-            BabyBear::from_wrapped_u32(747167305),
-            BabyBear::from_wrapped_u32(2968073607),
-            BabyBear::from_wrapped_u32(1053214930),
-            BabyBear::from_wrapped_u32(1074411832),
-            BabyBear::from_wrapped_u32(4016794508),
-            BabyBear::from_wrapped_u32(1570312929),
-            BabyBear::from_wrapped_u32(113576933),
-            BabyBear::from_wrapped_u32(4042581186),
-            BabyBear::from_wrapped_u32(3634515733),
-            BabyBear::from_wrapped_u32(1032701597),
-            BabyBear::from_wrapped_u32(2364839308),
-            BabyBear::from_wrapped_u32(3840286918),
-            BabyBear::from_wrapped_u32(888378655),
-            BabyBear::from_wrapped_u32(2520191583),
-        ],
-        [
-            BabyBear::from_wrapped_u32(36046858),
-            BabyBear::from_wrapped_u32(2927525953),
-            BabyBear::from_wrapped_u32(3912129105),
-            BabyBear::from_wrapped_u32(4004832531),
-            BabyBear::from_wrapped_u32(193772436),
-            BabyBear::from_wrapped_u32(1590247392),
-            BabyBear::from_wrapped_u32(4125818172),
-            BabyBear::from_wrapped_u32(2516251696),
-            BabyBear::from_wrapped_u32(4050945750),
-            BabyBear::from_wrapped_u32(269498914),
-            BabyBear::from_wrapped_u32(1973292656),
-            BabyBear::from_wrapped_u32(891403491),
-            BabyBear::from_wrapped_u32(1845429189),
-            BabyBear::from_wrapped_u32(2611996363),
-            BabyBear::from_wrapped_u32(2310542653),
-            BabyBear::from_wrapped_u32(4071195740),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3505307391),
-            BabyBear::from_wrapped_u32(786445290),
-            BabyBear::from_wrapped_u32(3815313971),
-            BabyBear::from_wrapped_u32(1111591756),
-            BabyBear::from_wrapped_u32(4233279834),
-            BabyBear::from_wrapped_u32(2775453034),
-            BabyBear::from_wrapped_u32(1991257625),
-            BabyBear::from_wrapped_u32(2940505809),
-            BabyBear::from_wrapped_u32(2751316206),
-            BabyBear::from_wrapped_u32(1028870679),
-            BabyBear::from_wrapped_u32(1282466273),
-            BabyBear::from_wrapped_u32(1059053371),
-            BabyBear::from_wrapped_u32(834521354),
-            BabyBear::from_wrapped_u32(138721483),
-            BabyBear::from_wrapped_u32(3100410803),
-            BabyBear::from_wrapped_u32(3843128331),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3878220780),
-            BabyBear::from_wrapped_u32(4058162439),
-            BabyBear::from_wrapped_u32(1478942487),
-            BabyBear::from_wrapped_u32(799012923),
-            BabyBear::from_wrapped_u32(496734827),
-            BabyBear::from_wrapped_u32(3521261236),
-            BabyBear::from_wrapped_u32(755421082),
-            BabyBear::from_wrapped_u32(1361409515),
-            BabyBear::from_wrapped_u32(392099473),
-            BabyBear::from_wrapped_u32(3178453393),
-            BabyBear::from_wrapped_u32(4068463721),
-            BabyBear::from_wrapped_u32(7935614),
-            BabyBear::from_wrapped_u32(4140885645),
-            BabyBear::from_wrapped_u32(2150748066),
-            BabyBear::from_wrapped_u32(1685210312),
-            BabyBear::from_wrapped_u32(3852983224),
-        ],
-        [
-            BabyBear::from_wrapped_u32(2896943075),
-            BabyBear::from_wrapped_u32(3087590927),
-            BabyBear::from_wrapped_u32(992175959),
-            BabyBear::from_wrapped_u32(970216228),
-            BabyBear::from_wrapped_u32(3473630090),
-            BabyBear::from_wrapped_u32(3899670400),
-            BabyBear::from_wrapped_u32(3603388822),
-            BabyBear::from_wrapped_u32(2633488197),
-            BabyBear::from_wrapped_u32(2479406964),
-            BabyBear::from_wrapped_u32(2420952999),
-            BabyBear::from_wrapped_u32(1852516800),
-            BabyBear::from_wrapped_u32(4253075697),
-            BabyBear::from_wrapped_u32(979699862),
-            BabyBear::from_wrapped_u32(1163403191),
-            BabyBear::from_wrapped_u32(1608599874),
-            BabyBear::from_wrapped_u32(3056104448),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3779109343),
-            BabyBear::from_wrapped_u32(536205958),
-            BabyBear::from_wrapped_u32(4183458361),
-            BabyBear::from_wrapped_u32(1649720295),
-            BabyBear::from_wrapped_u32(1444912244),
-            BabyBear::from_wrapped_u32(3122230878),
-            BabyBear::from_wrapped_u32(384301396),
-            BabyBear::from_wrapped_u32(4228198516),
-            BabyBear::from_wrapped_u32(1662916865),
-            BabyBear::from_wrapped_u32(4082161114),
-            BabyBear::from_wrapped_u32(2121897314),
-            BabyBear::from_wrapped_u32(1706239958),
-            BabyBear::from_wrapped_u32(4166959388),
-            BabyBear::from_wrapped_u32(1626054781),
-            BabyBear::from_wrapped_u32(3005858978),
-            BabyBear::from_wrapped_u32(1431907253),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1418914503),
-            BabyBear::from_wrapped_u32(1365856753),
-            BabyBear::from_wrapped_u32(3942715745),
-            BabyBear::from_wrapped_u32(1429155552),
-            BabyBear::from_wrapped_u32(3545642795),
-            BabyBear::from_wrapped_u32(3772474257),
-            BabyBear::from_wrapped_u32(1621094396),
-            BabyBear::from_wrapped_u32(2154399145),
-            BabyBear::from_wrapped_u32(826697382),
-            BabyBear::from_wrapped_u32(1700781391),
-            BabyBear::from_wrapped_u32(3539164324),
-            BabyBear::from_wrapped_u32(652815039),
-            BabyBear::from_wrapped_u32(442484755),
-            BabyBear::from_wrapped_u32(2055299391),
-            BabyBear::from_wrapped_u32(1064289978),
-            BabyBear::from_wrapped_u32(1152335780),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3417648695),
-            BabyBear::from_wrapped_u32(186040114),
-            BabyBear::from_wrapped_u32(3475580573),
-            BabyBear::from_wrapped_u32(2113941250),
-            BabyBear::from_wrapped_u32(1779573826),
-            BabyBear::from_wrapped_u32(1573808590),
-            BabyBear::from_wrapped_u32(3235694804),
-            BabyBear::from_wrapped_u32(2922195281),
-            BabyBear::from_wrapped_u32(1119462702),
-            BabyBear::from_wrapped_u32(3688305521),
-            BabyBear::from_wrapped_u32(1849567013),
-            BabyBear::from_wrapped_u32(667446787),
-            BabyBear::from_wrapped_u32(753897224),
-            BabyBear::from_wrapped_u32(1896396780),
-            BabyBear::from_wrapped_u32(3143026334),
-            BabyBear::from_wrapped_u32(3829603876),
-        ],
-        [
-            BabyBear::from_wrapped_u32(859661334),
-            BabyBear::from_wrapped_u32(3898844357),
-            BabyBear::from_wrapped_u32(180258337),
-            BabyBear::from_wrapped_u32(2321867017),
-            BabyBear::from_wrapped_u32(3599002504),
-            BabyBear::from_wrapped_u32(2886782421),
-            BabyBear::from_wrapped_u32(3038299378),
-            BabyBear::from_wrapped_u32(1035366250),
-            BabyBear::from_wrapped_u32(2038912197),
-            BabyBear::from_wrapped_u32(2920174523),
-            BabyBear::from_wrapped_u32(1277696101),
-            BabyBear::from_wrapped_u32(2785700290),
-            BabyBear::from_wrapped_u32(3806504335),
-            BabyBear::from_wrapped_u32(3518858933),
-            BabyBear::from_wrapped_u32(654843672),
-            BabyBear::from_wrapped_u32(2127120275),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1548195514),
-            BabyBear::from_wrapped_u32(2378056027),
-            BabyBear::from_wrapped_u32(390914568),
-            BabyBear::from_wrapped_u32(1472049779),
-            BabyBear::from_wrapped_u32(1552596765),
-            BabyBear::from_wrapped_u32(1905886441),
-            BabyBear::from_wrapped_u32(1611959354),
-            BabyBear::from_wrapped_u32(3653263304),
-            BabyBear::from_wrapped_u32(3423946386),
-            BabyBear::from_wrapped_u32(340857935),
-            BabyBear::from_wrapped_u32(2208879480),
-            BabyBear::from_wrapped_u32(139364268),
-            BabyBear::from_wrapped_u32(3447281773),
-            BabyBear::from_wrapped_u32(3777813707),
-            BabyBear::from_wrapped_u32(55640413),
-            BabyBear::from_wrapped_u32(4101901741),
-        ],
-        [
-            BabyBear::from_wrapped_u32(104929687),
-            BabyBear::from_wrapped_u32(1459980974),
-            BabyBear::from_wrapped_u32(1831234737),
-            BabyBear::from_wrapped_u32(457139004),
-            BabyBear::from_wrapped_u32(2581487628),
-            BabyBear::from_wrapped_u32(2112044563),
-            BabyBear::from_wrapped_u32(3567013861),
-            BabyBear::from_wrapped_u32(2792004347),
-            BabyBear::from_wrapped_u32(576325418),
-            BabyBear::from_wrapped_u32(41126132),
-            BabyBear::from_wrapped_u32(2713562324),
-            BabyBear::from_wrapped_u32(151213722),
-            BabyBear::from_wrapped_u32(2891185935),
-            BabyBear::from_wrapped_u32(546846420),
-            BabyBear::from_wrapped_u32(2939794919),
-            BabyBear::from_wrapped_u32(2543469905)
-        ],
-        [
-            BabyBear::from_wrapped_u32(2191909784),
-            BabyBear::from_wrapped_u32(3315138460),
-            BabyBear::from_wrapped_u32(530414574),
-            BabyBear::from_wrapped_u32(1242280418),
-            BabyBear::from_wrapped_u32(1211740715),
-            BabyBear::from_wrapped_u32(3993672165),
-            BabyBear::from_wrapped_u32(2505083323),
-            BabyBear::from_wrapped_u32(3845798801),
-            BabyBear::from_wrapped_u32(538768466),
-            BabyBear::from_wrapped_u32(2063567560),
-            BabyBear::from_wrapped_u32(3366148274),
-            BabyBear::from_wrapped_u32(1449831887),
-            BabyBear::from_wrapped_u32(2408012466),
-            BabyBear::from_wrapped_u32(294726285),
-            BabyBear::from_wrapped_u32(3943435493),
-            BabyBear::from_wrapped_u32(924016661),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3633138367),
-            BabyBear::from_wrapped_u32(3222789372),
-            BabyBear::from_wrapped_u32(809116305),
-            BabyBear::from_wrapped_u32(30100013),
-            BabyBear::from_wrapped_u32(2655172876),
-            BabyBear::from_wrapped_u32(2564247117),
-            BabyBear::from_wrapped_u32(2478649732),
-            BabyBear::from_wrapped_u32(4113689151),
-            BabyBear::from_wrapped_u32(4120146082),
-            BabyBear::from_wrapped_u32(2512308515),
-            BabyBear::from_wrapped_u32(650406041),
-            BabyBear::from_wrapped_u32(4240012393),
-            BabyBear::from_wrapped_u32(2683508708),
-            BabyBear::from_wrapped_u32(951073977),
-            BabyBear::from_wrapped_u32(3460081988),
-            BabyBear::from_wrapped_u32(339124269),
-        ],
-        [
-            BabyBear::from_wrapped_u32(130182653),
-            BabyBear::from_wrapped_u32(2755946749),
-            BabyBear::from_wrapped_u32(542600513),
-            BabyBear::from_wrapped_u32(2816103022),
-            BabyBear::from_wrapped_u32(1931786340),
-            BabyBear::from_wrapped_u32(2044470840),
-            BabyBear::from_wrapped_u32(1709908013),
-            BabyBear::from_wrapped_u32(2938369043),
-            BabyBear::from_wrapped_u32(3640399693),
-            BabyBear::from_wrapped_u32(1374470239),
-            BabyBear::from_wrapped_u32(2191149676),
-            BabyBear::from_wrapped_u32(2637495682),
-            BabyBear::from_wrapped_u32(4236394040),
-            BabyBear::from_wrapped_u32(2289358846),
-            BabyBear::from_wrapped_u32(3833368530),
-            BabyBear::from_wrapped_u32(974546524),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3306659113),
-            BabyBear::from_wrapped_u32(2234814261),
-            BabyBear::from_wrapped_u32(1188782305),
-            BabyBear::from_wrapped_u32(223782844),
-            BabyBear::from_wrapped_u32(2248980567),
-            BabyBear::from_wrapped_u32(2309786141),
-            BabyBear::from_wrapped_u32(2023401627),
-            BabyBear::from_wrapped_u32(3278877413),
-            BabyBear::from_wrapped_u32(2022138149),
-            BabyBear::from_wrapped_u32(575851471),
-            BabyBear::from_wrapped_u32(1612560780),
-            BabyBear::from_wrapped_u32(3926656936),
-            BabyBear::from_wrapped_u32(3318548977),
-            BabyBear::from_wrapped_u32(2591863678),
-            BabyBear::from_wrapped_u32(188109355),
-            BabyBear::from_wrapped_u32(4217723909),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1564209905),
-            BabyBear::from_wrapped_u32(2154197895),
-            BabyBear::from_wrapped_u32(2459687029),
-            BabyBear::from_wrapped_u32(2870634489),
-            BabyBear::from_wrapped_u32(1375012945),
-            BabyBear::from_wrapped_u32(1529454825),
-            BabyBear::from_wrapped_u32(306140690),
-            BabyBear::from_wrapped_u32(2855578299),
-            BabyBear::from_wrapped_u32(1246997295),
-            BabyBear::from_wrapped_u32(3024298763),
-            BabyBear::from_wrapped_u32(1915270363),
-            BabyBear::from_wrapped_u32(1218245412),
-            BabyBear::from_wrapped_u32(2479314020),
-            BabyBear::from_wrapped_u32(2989827755),
-            BabyBear::from_wrapped_u32(814378556),
-            BabyBear::from_wrapped_u32(4039775921),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1165280628),
-            BabyBear::from_wrapped_u32(1203983801),
-            BabyBear::from_wrapped_u32(3814740033),
-            BabyBear::from_wrapped_u32(1919627044),
-            BabyBear::from_wrapped_u32(600240215),
-            BabyBear::from_wrapped_u32(773269071),
-            BabyBear::from_wrapped_u32(486685186),
-            BabyBear::from_wrapped_u32(4254048810),
-            BabyBear::from_wrapped_u32(1415023565),
-            BabyBear::from_wrapped_u32(502840102),
-            BabyBear::from_wrapped_u32(4225648358),
-            BabyBear::from_wrapped_u32(510217063),
-            BabyBear::from_wrapped_u32(166444818),
-            BabyBear::from_wrapped_u32(1430745893),
-            BabyBear::from_wrapped_u32(1376516190),
-            BabyBear::from_wrapped_u32(1775891321),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1170945922),
-            BabyBear::from_wrapped_u32(1105391877),
-            BabyBear::from_wrapped_u32(261536467),
-            BabyBear::from_wrapped_u32(1401687994),
-            BabyBear::from_wrapped_u32(1022529847),
-            BabyBear::from_wrapped_u32(2476446456),
-            BabyBear::from_wrapped_u32(2603844878),
-            BabyBear::from_wrapped_u32(3706336043),
-            BabyBear::from_wrapped_u32(3463053714),
-            BabyBear::from_wrapped_u32(1509644517),
-            BabyBear::from_wrapped_u32(588552318),
-            BabyBear::from_wrapped_u32(65252581),
-            BabyBear::from_wrapped_u32(3696502656),
-            BabyBear::from_wrapped_u32(2183330763),
-            BabyBear::from_wrapped_u32(3664021233),
-            BabyBear::from_wrapped_u32(1643809916),
-        ],
-        [
-            BabyBear::from_wrapped_u32(2922875898),
-            BabyBear::from_wrapped_u32(3740690643),
-            BabyBear::from_wrapped_u32(3932461140),
-            BabyBear::from_wrapped_u32(161156271),
-            BabyBear::from_wrapped_u32(2619943483),
-            BabyBear::from_wrapped_u32(4077039509),
-            BabyBear::from_wrapped_u32(2921201703),
-            BabyBear::from_wrapped_u32(2085619718),
-            BabyBear::from_wrapped_u32(2065264646),
-            BabyBear::from_wrapped_u32(2615693812),
-            BabyBear::from_wrapped_u32(3116555433),
-            BabyBear::from_wrapped_u32(246100007),
-            BabyBear::from_wrapped_u32(4281387154),
-            BabyBear::from_wrapped_u32(4046141001),
-            BabyBear::from_wrapped_u32(4027749321),
-            BabyBear::from_wrapped_u32(111611860),
-        ],
-        [
-            BabyBear::from_wrapped_u32(2066954820),
-            BabyBear::from_wrapped_u32(2502099969),
-            BabyBear::from_wrapped_u32(2915053115),
-            BabyBear::from_wrapped_u32(2362518586),
-            BabyBear::from_wrapped_u32(366091708),
-            BabyBear::from_wrapped_u32(2083204932),
-            BabyBear::from_wrapped_u32(4138385632),
-            BabyBear::from_wrapped_u32(3195157567),
-            BabyBear::from_wrapped_u32(1318086382),
-            BabyBear::from_wrapped_u32(521723799),
-            BabyBear::from_wrapped_u32(702443405),
-            BabyBear::from_wrapped_u32(2507670985),
-            BabyBear::from_wrapped_u32(1760347557),
-            BabyBear::from_wrapped_u32(2631999893),
-            BabyBear::from_wrapped_u32(1672737554),
-            BabyBear::from_wrapped_u32(1060867760),
-        ],
-        [
-            BabyBear::from_wrapped_u32(2359801781),
-            BabyBear::from_wrapped_u32(2800231467),
-            BabyBear::from_wrapped_u32(3010357035),
-            BabyBear::from_wrapped_u32(1035997899),
-            BabyBear::from_wrapped_u32(1210110952),
-            BabyBear::from_wrapped_u32(1018506770),
-            BabyBear::from_wrapped_u32(2799468177),
-            BabyBear::from_wrapped_u32(1479380761),
-            BabyBear::from_wrapped_u32(1536021911),
-            BabyBear::from_wrapped_u32(358993854),
-            BabyBear::from_wrapped_u32(579904113),
-            BabyBear::from_wrapped_u32(3432144800),
-            BabyBear::from_wrapped_u32(3625515809),
-            BabyBear::from_wrapped_u32(199241497),
-            BabyBear::from_wrapped_u32(4058304109),
-            BabyBear::from_wrapped_u32(2590164234),
-        ],
-        [
-            BabyBear::from_wrapped_u32(1688530738),
-            BabyBear::from_wrapped_u32(1580733335),
-            BabyBear::from_wrapped_u32(2443981517),
-            BabyBear::from_wrapped_u32(2206270565),
-            BabyBear::from_wrapped_u32(2780074229),
-            BabyBear::from_wrapped_u32(2628739677),
-            BabyBear::from_wrapped_u32(2940123659),
-            BabyBear::from_wrapped_u32(4145206827),
-            BabyBear::from_wrapped_u32(3572278009),
-            BabyBear::from_wrapped_u32(2779607509),
-            BabyBear::from_wrapped_u32(1098718697),
-            BabyBear::from_wrapped_u32(1424913749),
-            BabyBear::from_wrapped_u32(2224415875),
-            BabyBear::from_wrapped_u32(1108922178),
-            BabyBear::from_wrapped_u32(3646272562),
-            BabyBear::from_wrapped_u32(3935186184),
-        ],
-        [
-            BabyBear::from_wrapped_u32(820046587),
-            BabyBear::from_wrapped_u32(1393386250),
-            BabyBear::from_wrapped_u32(2665818575),
-            BabyBear::from_wrapped_u32(2231782019),
-            BabyBear::from_wrapped_u32(672377010),
-            BabyBear::from_wrapped_u32(1920315467),
-            BabyBear::from_wrapped_u32(1913164407),
-            BabyBear::from_wrapped_u32(2029526876),
-            BabyBear::from_wrapped_u32(2629271820),
-            BabyBear::from_wrapped_u32(384320012),
-            BabyBear::from_wrapped_u32(4112320585),
-            BabyBear::from_wrapped_u32(3131824773),
-            BabyBear::from_wrapped_u32(2347818197),
-            BabyBear::from_wrapped_u32(2220997386),
-            BabyBear::from_wrapped_u32(1772368609),
-            BabyBear::from_wrapped_u32(2579960095),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3544930873),
-            BabyBear::from_wrapped_u32(225847443),
-            BabyBear::from_wrapped_u32(3070082278),
-            BabyBear::from_wrapped_u32(95643305),
-            BabyBear::from_wrapped_u32(3438572042),
-            BabyBear::from_wrapped_u32(3312856509),
-            BabyBear::from_wrapped_u32(615850007),
-            BabyBear::from_wrapped_u32(1863868773),
-            BabyBear::from_wrapped_u32(803582265),
-            BabyBear::from_wrapped_u32(3461976859),
-            BabyBear::from_wrapped_u32(2903025799),
-            BabyBear::from_wrapped_u32(1482092434),
-            BabyBear::from_wrapped_u32(3902972499),
-            BabyBear::from_wrapped_u32(3872341868),
-            BabyBear::from_wrapped_u32(1530411808),
-            BabyBear::from_wrapped_u32(2214923584),
-        ],
-        [
-            BabyBear::from_wrapped_u32(3118792481),
-            BabyBear::from_wrapped_u32(2241076515),
-            BabyBear::from_wrapped_u32(3983669831),
-            BabyBear::from_wrapped_u32(3180915147),
-            BabyBear::from_wrapped_u32(3838626501),
-            BabyBear::from_wrapped_u32(1921630011),
-            BabyBear::from_wrapped_u32(3415351771),
-            BabyBear::from_wrapped_u32(2249953859),
-            BabyBear::from_wrapped_u32(3755081630),
-            BabyBear::from_wrapped_u32(486327260),
-            BabyBear::from_wrapped_u32(1227575720),
-            BabyBear::from_wrapped_u32(3643869379),
-            BabyBear::from_wrapped_u32(2982026073),
-            BabyBear::from_wrapped_u32(2466043731),
-            BabyBear::from_wrapped_u32(1982634375),
-            BabyBear::from_wrapped_u32(3769609014),
-        ],
-        [
-            BabyBear::from_wrapped_u32(2195455495),
-            BabyBear::from_wrapped_u32(2596863283),
-            BabyBear::from_wrapped_u32(4244994973),
-            BabyBear::from_wrapped_u32(1983609348),
-            BabyBear::from_wrapped_u32(4019674395),
-            BabyBear::from_wrapped_u32(3469982031),
-            BabyBear::from_wrapped_u32(1458697570),
-            BabyBear::from_wrapped_u32(1593516217),
-            BabyBear::from_wrapped_u32(1963896497),
-            BabyBear::from_wrapped_u32(3115309118),
-            BabyBear::from_wrapped_u32(1659132465),
-            BabyBear::from_wrapped_u32(2536770756),
-            BabyBear::from_wrapped_u32(3059294171),
-            BabyBear::from_wrapped_u32(2618031334),
-            BabyBear::from_wrapped_u32(2040903247),
-            BabyBear::from_wrapped_u32(3799795076),
-        ]
-    ];
-
-
-    pub static ref RC_16_30_U32: [[u32; 16]; 30] = [
-        [
-            (2110014213),
-            (3964964605),
-            (2190662774),
-            (2732996483),
-            (640767983),
-            (3403899136),
-            (1716033721),
-            (1606702601),
-            (3759873288),
-            (1466015491),
-            (1498308946),
-            (2844375094),
-            (3042463841),
-            (1969905919),
-            (4109944726),
-            (3925048366),
-        ],
-        [
-            (3706859504),
-            (759122502),
-            (3167665446),
-            (1131812921),
-            (1080754908),
-            (4080114493),
-            (893583089),
-            (2019677373),
-            (3128604556),
-            (580640471),
-            (3277620260),
-            (842931656),
-            (548879852),
-            (3608554714),
-            (3575647916),
-            (81826002),
-        ],
-        [
-            (4289086263),
-            (1563933798),
-            (1440025885),
-            (184445025),
-            (2598651360),
-            (1396647410),
-            (1575877922),
-            (3303853401),
-            (137125468),
-            (765010148),
-            (633675867),
-            (2037803363),
-            (2573389828),
-            (1895729703),
-            (541515871),
-            (1783382863),
-        ],
-        [
-            (2641856484),
-            (3035743342),
-            (3672796326),
-            (245668751),
-            (2025460432),
-            (201609705),
-            (286217151),
-            (4093475563),
-            (2519572182),
-            (3080699870),
-            (2762001832),
-            (1244250808),
-            (606038199),
-            (3182740831),
-            (73007766),
-            (2572204153),
-        ],
-        [
-            (1196780786),
-            (3447394443),
-            (747167305),
-            (2968073607),
-            (1053214930),
-            (1074411832),
-            (4016794508),
-            (1570312929),
-            (113576933),
-            (4042581186),
-            (3634515733),
-            (1032701597),
-            (2364839308),
-            (3840286918),
-            (888378655),
-            (2520191583),
-        ],
-        [
-            (36046858),
-            (2927525953),
-            (3912129105),
-            (4004832531),
-            (193772436),
-            (1590247392),
-            (4125818172),
-            (2516251696),
-            (4050945750),
-            (269498914),
-            (1973292656),
-            (891403491),
-            (1845429189),
-            (2611996363),
-            (2310542653),
-            (4071195740),
-        ],
-        [
-            (3505307391),
-            (786445290),
-            (3815313971),
-            (1111591756),
-            (4233279834),
-            (2775453034),
-            (1991257625),
-            (2940505809),
-            (2751316206),
-            (1028870679),
-            (1282466273),
-            (1059053371),
-            (834521354),
-            (138721483),
-            (3100410803),
-            (3843128331),
-        ],
-        [
-            (3878220780),
-            (4058162439),
-            (1478942487),
-            (799012923),
-            (496734827),
-            (3521261236),
-            (755421082),
-            (1361409515),
-            (392099473),
-            (3178453393),
-            (4068463721),
-            (7935614),
-            (4140885645),
-            (2150748066),
-            (1685210312),
-            (3852983224),
-        ],
-        [
-            (2896943075),
-            (3087590927),
-            (992175959),
-            (970216228),
-            (3473630090),
-            (3899670400),
-            (3603388822),
-            (2633488197),
-            (2479406964),
-            (2420952999),
-            (1852516800),
-            (4253075697),
-            (979699862),
-            (1163403191),
-            (1608599874),
-            (3056104448),
-        ],
-        [
-            (3779109343),
-            (536205958),
-            (4183458361),
-            (1649720295),
-            (1444912244),
-            (3122230878),
-            (384301396),
-            (4228198516),
-            (1662916865),
-            (4082161114),
-            (2121897314),
-            (1706239958),
-            (4166959388),
-            (1626054781),
-            (3005858978),
-            (1431907253),
-        ],
-        [
-            (1418914503),
-            (1365856753),
-            (3942715745),
-            (1429155552),
-            (3545642795),
-            (3772474257),
-            (1621094396),
-            (2154399145),
-            (826697382),
-            (1700781391),
-            (3539164324),
-            (652815039),
-            (442484755),
-            (2055299391),
-            (1064289978),
-            (1152335780),
-        ],
-        [
-            (3417648695),
-            (186040114),
-            (3475580573),
-            (2113941250),
-            (1779573826),
-            (1573808590),
-            (3235694804),
-            (2922195281),
-            (1119462702),
-            (3688305521),
-            (1849567013),
-            (667446787),
-            (753897224),
-            (1896396780),
-            (3143026334),
-            (3829603876),
-        ],
-        [
-            (859661334),
-            (3898844357),
-            (180258337),
-            (2321867017),
-            (3599002504),
-            (2886782421),
-            (3038299378),
-            (1035366250),
-            (2038912197),
-            (2920174523),
-            (1277696101),
-            (2785700290),
-            (3806504335),
-            (3518858933),
-            (654843672),
-            (2127120275),
-        ],
-        [
-            (1548195514),
-            (2378056027),
-            (390914568),
-            (1472049779),
-            (1552596765),
-            (1905886441),
-            (1611959354),
-            (3653263304),
-            (3423946386),
-            (340857935),
-            (2208879480),
-            (139364268),
-            (3447281773),
-            (3777813707),
-            (55640413),
-            (4101901741),
-        ],
-        [
-            (104929687),
-            (1459980974),
-            (1831234737),
-            (457139004),
-            (2581487628),
-            (2112044563),
-            (3567013861),
-            (2792004347),
-            (576325418),
-            (41126132),
-            (2713562324),
-            (151213722),
-            (2891185935),
-            (546846420),
-            (2939794919),
-            (2543469905)
-        ],
-        [
-            (2191909784),
-            (3315138460),
-            (530414574),
-            (1242280418),
-            (1211740715),
-            (3993672165),
-            (2505083323),
-            (3845798801),
-            (538768466),
-            (2063567560),
-            (3366148274),
-            (1449831887),
-            (2408012466),
-            (294726285),
-            (3943435493),
-            (924016661),
-        ],
-        [
-            (3633138367),
-            (3222789372),
-            (809116305),
-            (30100013),
-            (2655172876),
-            (2564247117),
-            (2478649732),
-            (4113689151),
-            (4120146082),
-            (2512308515),
-            (650406041),
-            (4240012393),
-            (2683508708),
-            (951073977),
-            (3460081988),
-            (339124269),
-        ],
-        [
-            (130182653),
-            (2755946749),
-            (542600513),
-            (2816103022),
-            (1931786340),
-            (2044470840),
-            (1709908013),
-            (2938369043),
-            (3640399693),
-            (1374470239),
-            (2191149676),
-            (2637495682),
-            (4236394040),
-            (2289358846),
-            (3833368530),
-            (974546524),
-        ],
-        [
-            (3306659113),
-            (2234814261),
-            (1188782305),
-            (223782844),
-            (2248980567),
-            (2309786141),
-            (2023401627),
-            (3278877413),
-            (2022138149),
-            (575851471),
-            (1612560780),
-            (3926656936),
-            (3318548977),
-            (2591863678),
-            (188109355),
-            (4217723909),
-        ],
-        [
-            (1564209905),
-            (2154197895),
-            (2459687029),
-            (2870634489),
-            (1375012945),
-            (1529454825),
-            (306140690),
-            (2855578299),
-            (1246997295),
-            (3024298763),
-            (1915270363),
-            (1218245412),
-            (2479314020),
-            (2989827755),
-            (814378556),
-            (4039775921),
-        ],
-        [
-            (1165280628),
-            (1203983801),
-            (3814740033),
-            (1919627044),
-            (600240215),
-            (773269071),
-            (486685186),
-            (4254048810),
-            (1415023565),
-            (502840102),
-            (4225648358),
-            (510217063),
-            (166444818),
-            (1430745893),
-            (1376516190),
-            (1775891321),
-        ],
-        [
-            (1170945922),
-            (1105391877),
-            (261536467),
-            (1401687994),
-            (1022529847),
-            (2476446456),
-            (2603844878),
-            (3706336043),
-            (3463053714),
-            (1509644517),
-            (588552318),
-            (65252581),
-            (3696502656),
-            (2183330763),
-            (3664021233),
-            (1643809916),
-        ],
-        [
-            (2922875898),
-            (3740690643),
-            (3932461140),
-            (161156271),
-            (2619943483),
-            (4077039509),
-            (2921201703),
-            (2085619718),
-            (2065264646),
-            (2615693812),
-            (3116555433),
-            (246100007),
-            (4281387154),
-            (4046141001),
-            (4027749321),
-            (111611860),
-        ],
-        [
-            (2066954820),
-            (2502099969),
-            (2915053115),
-            (2362518586),
-            (366091708),
-            (2083204932),
-            (4138385632),
-            (3195157567),
-            (1318086382),
-            (521723799),
-            (702443405),
-            (2507670985),
-            (1760347557),
-            (2631999893),
-            (1672737554),
-            (1060867760),
-        ],
-        [
-            (2359801781),
-            (2800231467),
-            (3010357035),
-            (1035997899),
-            (1210110952),
-            (1018506770),
-            (2799468177),
-            (1479380761),
-            (1536021911),
-            (358993854),
-            (579904113),
-            (3432144800),
-            (3625515809),
-            (199241497),
-            (4058304109),
-            (2590164234),
-        ],
-        [
-            (1688530738),
-            (1580733335),
-            (2443981517),
-            (2206270565),
-            (2780074229),
-            (2628739677),
-            (2940123659),
-            (4145206827),
-            (3572278009),
-            (2779607509),
-            (1098718697),
-            (1424913749),
-            (2224415875),
-            (1108922178),
-            (3646272562),
-            (3935186184),
-        ],
-        [
-            (820046587),
-            (1393386250),
-            (2665818575),
-            (2231782019),
-            (672377010),
-            (1920315467),
-            (1913164407),
-            (2029526876),
-            (2629271820),
-            (384320012),
-            (4112320585),
-            (3131824773),
-            (2347818197),
-            (2220997386),
-            (1772368609),
-            (2579960095),
-        ],
-        [
-            (3544930873),
-            (225847443),
-            (3070082278),
-            (95643305),
-            (3438572042),
-            (3312856509),
-            (615850007),
-            (1863868773),
-            (803582265),
-            (3461976859),
-            (2903025799),
-            (1482092434),
-            (3902972499),
-            (3872341868),
-            (1530411808),
-            (2214923584),
-        ],
-        [
-            (3118792481),
-            (2241076515),
-            (3983669831),
-            (3180915147),
-            (3838626501),
-            (1921630011),
-            (3415351771),
-            (2249953859),
-            (3755081630),
-            (486327260),
-            (1227575720),
-            (3643869379),
-            (2982026073),
-            (2466043731),
-            (1982634375),
-            (3769609014),
-        ],
-        [
-            (2195455495),
-            (2596863283),
-            (4244994973),
-            (1983609348),
-            (4019674395),
-            (3469982031),
-            (1458697570),
-            (1593516217),
-            (1963896497),
-            (3115309118),
-            (1659132465),
-            (2536770756),
-            (3059294171),
-            (2618031334),
-            (2040903247),
-            (3799795076),
-        ]
-    ];
-}
-
-pub fn poseidon2_init(
-) -> Poseidon2<BabyBear, Poseidon2ExternalMatrixGeneral, DiffusionMatrixBabyBear, 16, 7> {
-    const ROUNDS_F: usize = 8;
-    const ROUNDS_P: usize = 13;
-    let mut round_constants = RC_16_30.to_vec();
-    let internal_start = ROUNDS_F / 2;
-    let internal_end = (ROUNDS_F / 2) + ROUNDS_P;
-    let internal_round_constants =
-        round_constants.drain(internal_start..internal_end).map(|vec| vec[0]).collect::<Vec<_>>();
-    let external_round_constants = round_constants;
-    Poseidon2::new(
-        ROUNDS_F,
-        external_round_constants,
-        Poseidon2ExternalMatrixGeneral,
-        ROUNDS_P,
-        internal_round_constants,
-        DiffusionMatrixBabyBear,
-    )
-}
-
-use p3_symmetric::{CryptographicHasher, PaddingFreeSponge};
-
-pub fn poseidon2_hash(input: Vec<BabyBear>) -> [BabyBear; 8] {
-    POSEIDON2_HASHER.hash_iter(input)
-}
-
-pub fn poseidon2_hasher() -> PaddingFreeSponge<
-    Poseidon2<BabyBear, Poseidon2ExternalMatrixGeneral, DiffusionMatrixBabyBear, 16, 7>,
-    16,
-    8,
-    8,
-> {
-    let hasher = poseidon2_init();
-    PaddingFreeSponge::<
-        Poseidon2<BabyBear, Poseidon2ExternalMatrixGeneral, DiffusionMatrixBabyBear, 16, 7>,
-        16,
-        8,
-        8,
-    >::new(hasher)
-}
-
-lazy_static! {
-    pub static ref POSEIDON2_HASHER: PaddingFreeSponge::<
-        Poseidon2<BabyBear, Poseidon2ExternalMatrixGeneral, DiffusionMatrixBabyBear, 16, 7>,
-        16,
-        8,
-        8,
-    > = poseidon2_hasher();
-}
-
-/// Append a single deferred proof to a hash chain of deferred proofs.
-pub fn hash_deferred_proof(
-    prev_digest: &[BabyBear; 8],
-    vk_digest: &[BabyBear; 8],
-    pv_digest: &[BabyBear; 32],
-) -> [BabyBear; 8] {
-    let mut inputs = Vec::with_capacity(48);
-    inputs.extend_from_slice(prev_digest);
-    inputs.extend_from_slice(vk_digest);
-    inputs.extend_from_slice(pv_digest);
-    poseidon2_hash(inputs.to_vec())
-}
+pub use consts::*;
+pub use io::*;
+pub use poseidon2::*;
+pub use types::*;
