+++ crates/zkvm/entrypoint/src/syscalls/mod.rs
@@ -7,6 +7,7 @@ mod halt;
 mod io;
 mod keccak_permute;
 mod memory;
+mod poseidon2_permute;
 mod secp256k1;
 mod secp256r1;
 mod sha_compress;
@@ -26,6 +27,7 @@ pub use halt::*;
 pub use io::*;
 pub use keccak_permute::*;
 pub use memory::*;
+pub use poseidon2_permute::*;
 pub use secp256k1::*;
 pub use secp256r1::*;
 pub use sha_compress::*;
@@ -152,3 +154,6 @@ pub const BN254_FP2_SUB: u32 = 0x00_01_01_2A;
 
 /// Executes the `BN254_FP2_MUL` precompile.
 pub const BN254_FP2_MUL: u32 = 0x00_01_01_2B;
+
+/// Executes the `POSEIDON2_PERMUTE` precompile.
+pub const POSEIDON2_PERMUTE: u32 = 0x00_01_01_2F;
